+++ src/test/converter2/issues/gh2364.ts
@@ -8,3 +8,12 @@ export let NS = {
 export namespace NS {
     export type T = string;
 }
+
+export type NS2 = { foo: string };
+
+/**
+ * @namespace
+ */
+export const NS2 = {
+    property: "blah",
+};